refactor: adjust components name
