+++ src/components/FileInput/styled.ts
